Resume Parser/requirements.txt
spacy
python-docx
scikit-learn
fuzzywuzzy
